refactor(application): split commands/mod.rs into per-command files
application/commands/mod.rs held every app-registry command
(list/info/init/validate/update/lifecycle/release/add/deploy) in one
2272-line file. Split into one file per command, with deploy/ and
add_extension/ as subdirectories for the larger nested groups, plus a
shared schemas.rs for the output_schema! definitions.
