+++ rust/src/application/commands/mod.rs
@@ -1,177 +1,41 @@
-use cli_engine::{
-    CommandResult, CommandSpec, GroupSpec, NextAction, NextActionParam, RuntimeCommandSpec,
-    RuntimeGroupSpec, StreamSender, TableColumn, Tier,
-};
-use serde_json::{Value, json};
+//! `gddy platform app` — register, configure, and deploy GoDaddy
+//! developer-platform applications described by a godaddy.toml manifest.
 
-use crate::application::client::{ApplicationClient, UploadOptions, api_url_for_env};
+use cli_engine::{GroupSpec, NextAction, NextActionParam, RuntimeGroupSpec};
+
+use crate::application::client::api_url_for_env;
 use crate::next_action::{next_action, required_value};
-use crate::output_schema::output_schema;
-// App-registry mutations declare their scopes so `apps.app-registry:write` is
-// requested on demand (OAuth step-up), not granted at every login — it's a
-// rarely-used operation for most customers.
-use crate::scopes::{APP_REGISTRY_READ, APP_REGISTRY_WRITE};
 
-output_schema!(ApplicationSummary {
-    "id": "string";
-    "name": "string";
-    "label": "string", optional;
-    "description": "string", optional;
-    "status": "string";
-    "url": "string", optional;
-    "proxyUrl": "string", optional;
-});
+mod add;
+mod add_extension;
+mod deploy;
+mod info;
+mod init;
+mod lifecycle;
+mod list;
+mod release;
+mod schemas;
+mod update;
+mod validate;
 
-output_schema!(ApplicationInit {
-    "id": "string";
-    "name": "string";
-    "status": "string";
-    "clientId": "string";
-    "orgId": "string";
-    "url": "string";
-    "proxyUrl": "string";
-    "authorizationScopes": "[]string";
-    "oauthGrantTypes": "[]string";
-    "filesWritten": "object";
-});
-
-output_schema!(ApplicationUpdate {
-    "id": "string";
-    "clientId": "string";
-    "name": "string";
-    "label": "string", optional;
-    "description": "string", optional;
-    "status": "string";
-    "url": "string", optional;
-    "proxyUrl": "string", optional;
-    "authorizationScopes": "[]string";
-});
-
-output_schema!(ApplicationRef {
-    "id": "string";
-});
-
-output_schema!(ApplicationArchive {
-    "id": "string";
-    "name": "string";
-    "label": "string", optional;
-    "status": "string";
-    "createdAt": "string";
-    "archivedAt": "string";
-});
-
-output_schema!(ApplicationRelease {
-    "id": "string";
-    "version": "string";
-    "description": "string", optional;
-    "createdAt": "string";
-});
-
-output_schema!(ValidationResult {
-    "valid": "bool";
-    "errors": "[]string";
-    "warnings": "[]string";
-});
-
-output_schema!(ConfigAction {
-    "name": "string";
-    "url": "string";
-});
-
-output_schema!(ConfigSubscription {
-    "name": "string";
-    "url": "string";
-    "events": "[]string";
-});
-
-output_schema!(ExtensionHandle {
-    "name": "string";
-    "handle": "string";
-    "type": "string";
-});
-
-output_schema!(ExtensionBlocks {
-    "source": "string";
-    "type": "string";
-});
-
-async fn make_client(ctx: &cli_engine::CommandContext) -> cli_engine::Result<ApplicationClient> {
+async fn make_client(
+    ctx: &cli_engine::CommandContext,
+) -> cli_engine::Result<crate::application::client::ApplicationClient> {
     // Lazily resolve the credential; this triggers the auth flow only for
     // commands that actually call the API.
     let token = ctx.credential().await?.token;
     let base_url = api_url_for_env(&ctx.middleware.env)?;
-    Ok(ApplicationClient::new(base_url, token))
+    Ok(crate::application::client::ApplicationClient::new(
+        base_url, token,
+    ))
 }
 
 fn client_err(e: crate::application::client::ClientError) -> cli_engine::CliCoreError {
     crate::error::GddyError::from(e).into_cli_error()
 }
 
-/// Builds the terminal `{"type":"error",...}` event for a streaming command,
-/// reusing `cli_engine::build_error_envelope` so the `code`/`message`/`fix`
-/// match what a non-streaming command would have rendered for the same error.
-fn deploy_error_event(err: &cli_engine::CliCoreError) -> Value {
-    let envelope = cli_engine::build_error_envelope(err, "applications");
-    // `build_error_envelope` always populates `.error` with a non-empty code;
-    // this fallback only guards against a future change to that guarantee.
-    let (code, message) = envelope
-        .error
-        .map(|e| (e.code, e.message))
-        .unwrap_or_else(|| ("ERROR".to_owned(), err.to_string()));
-    let mut event = json!({
-        "type": "error",
-        "ok": false,
-        "error": { "code": code, "message": message },
-        "next_actions": [],
-    });
-    if let Some(fix) = envelope.fix.filter(|f| !f.is_empty()) {
-        event["fix"] = json!(fix);
-    }
-    event
-}
-
-/// Emit the terminal error event on a streaming command, then return the
-/// error so the handler can still fail the run via `?`.
-async fn fail_deploy(
-    sender: &StreamSender,
-    err: cli_engine::CliCoreError,
-) -> cli_engine::CliCoreError {
-    sender.send(deploy_error_event(&err)).await;
-    err
-}
-
-/// Builds the terminal `{"type":"result",...}` event for a successful
-/// `platform app deploy` run.
-fn deploy_result_event(
-    name: &str,
-    application_id: &str,
-    release_id: &str,
-    extensions: usize,
-) -> Value {
-    json!({
-        "type": "result",
-        "ok": true,
-        "result": {
-            "application": name,
-            "applicationId": application_id,
-            "releaseId": release_id,
-            "extensions": extensions,
-            "status": "ACTIVE",
-        },
-        "next_actions": deploy_next_actions(name),
-    })
-}
-
-/// Wrap a fallible step: on `Err`, emit the terminal error event before
-/// propagating, so every failure path produces exactly one terminal line.
-async fn tap_deploy_err<T>(
-    sender: &StreamSender,
-    result: cli_engine::Result<T>,
-) -> cli_engine::Result<T> {
-    match result {
-        Ok(v) => Ok(v),
-        Err(e) => Err(fail_deploy(sender, e).await),
-    }
+fn validation_err(message: impl Into<String>) -> cli_engine::CliCoreError {
+    crate::error::GddyError::validation(message).into_cli_error()
 }
 
 /// Next-actions after mutating local godaddy.toml (add action/subscription/extension).
@@ -196,25 +60,6 @@ fn add_config_next_actions(app_name: &str) -> Vec<NextAction> {
     ]
 }
 
-/// Next-actions after a successful deploy.
-fn deploy_next_actions(name: &str) -> Vec<NextAction> {
-    vec![
-        next_action(
-            "platform app enable <name> --store-id <store-id>",
-            "Enable the application on a store",
-        )
-        .with_param("name", required_value(name))
-        .with_param("store-id", NextActionParam::required()),
-        next_action(
-            "platform app info --name <name>",
-            "Inspect deployment status",
-        )
-        .with_param("name", required_value(name)),
-        next_action("platform app deploy --name <name>", "Rerun deployment")
-            .with_param("name", required_value(name)),
-    ]
-}
-
 pub fn application_group() -> RuntimeGroupSpec {
     RuntimeGroupSpec::new(
         GroupSpec::new("app", "Manage GoDaddy Platform apps")
@@ -227,2046 +72,32 @@ pub fn application_group() -> RuntimeGroupSpec {
             )
             .with_alias("application"),
     )
-    .with_command(list_command())
-    .with_command(info_command())
-    .with_command(init_command())
-    .with_command(validate_command())
-    .with_command(update_command())
-    .with_command(enable_command())
-    .with_command(disable_command())
-    .with_command(archive_command())
-    .with_command(release_command())
-    .with_command(deploy_command())
-    .with_group(add_group())
-}
-
-fn list_command() -> RuntimeCommandSpec {
-    RuntimeCommandSpec::new_with_context(
-        CommandSpec::new("list", "List all applications")
-            .with_long(
-                "List all GoDaddy developer-platform applications visible to the current \
-                account. Use `gddy platform app info --name <name>` to fetch full details for a \
-                single application.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Read)
-            .with_default_fields("name,label,status")
-            .with_output_schema::<ApplicationSummary>(),
-        |ctx| async move {
-            let client = make_client(&ctx).await?;
-            let data = client.list_applications().await.map_err(client_err)?;
-            Ok(CommandResult::new(data).with_next_actions(vec![
-                next_action(
-                    "platform app info --name <name>",
-                    "Get details for a specific application",
-                )
-                .with_param("name", NextActionParam::required()),
-                next_action(
-                    "platform app init --name <name> --description <description> --url <url> --proxy-url <proxy-url> --scopes <scopes>",
-                    "Initialize a new application",
-                ),
-            ]))
-        },
-    )
-}
-
-#[derive(Debug, Clone, clap::Args)]
-struct InfoArgs {
-    /// Application name.
-    #[arg(long, short = 'n', value_name = "NAME")]
-    name: String,
-}
-
-fn info_command() -> RuntimeCommandSpec {
-    RuntimeCommandSpec::new_typed_with_context::<InfoArgs, _, _, _>(
-        CommandSpec::from_args::<InfoArgs>("info", "Get application details")
-            .with_long(
-                "Fetch full details for a single GoDaddy developer-platform application by \
-                name, including status, URLs, and authorization scopes. Use `gddy platform app \
-                list` \
-                to find available names.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Read)
-            .with_output_schema::<ApplicationSummary>(),
-        |ctx, args: InfoArgs| async move {
-            let name = args.name;
-            let client = make_client(&ctx).await?;
-            let data = client.get_application(&name).await.map_err(client_err)?;
-            let app = &data["application"];
-            if app.is_null() {
-                return Err(crate::error::GddyError::not_found(format!(
-                    "application '{name}' not found"
-                ))
-                .into_cli_error());
-            }
-            let app_id = app["id"].as_str().unwrap_or("").to_owned();
-            Ok(CommandResult::new(app.clone()).with_next_actions(vec![
-                next_action(
-                    "platform app validate <name>",
-                    "Validate application configuration",
-                )
-                .with_param("name", required_value(&name)),
-                next_action(
-                    "platform app update --id <id> [--label <label>] [--description <description>] [--status <status>]",
-                    "Update application configuration",
-                )
-                .with_param("id", required_value(&app_id))
-                .with_param(
-                    "status",
-                    NextActionParam {
-                        r#enum: vec!["ACTIVE".to_owned(), "INACTIVE".to_owned()],
-                        ..Default::default()
-                    },
-                ),
-                next_action(
-                    "platform app release --application-id <application-id> --version <version>",
-                    "Create a release",
-                )
-                .with_param("application-id", required_value(&app_id))
-                .with_param("version", NextActionParam::required()),
-                next_action(
-                    "platform app deploy --name <name>",
-                    "Deploy this application",
-                )
-                .with_param("name", required_value(&name)),
-            ]))
-        },
-    )
-}
-
-#[derive(Debug, Clone, clap::Args)]
-struct InitArgs {
-    /// Application name (used as label if --label is not set).
-    #[arg(long, short = 'n', value_name = "NAME")]
-    name: Option<String>,
-
-    /// Display label (defaults to name).
-    #[arg(long, value_name = "LABEL")]
-    label: Option<String>,
-
-    /// Application description.
-    #[arg(long, value_name = "TEXT")]
-    description: Option<String>,
-
-    /// Application URL (must be public HTTP(S)).
-    #[arg(long, value_name = "URL")]
-    url: Option<String>,
-
-    /// Proxy URL (must be public HTTP(S)).
-    #[arg(long, value_name = "URL")]
-    proxy_url: Option<String>,
-
-    /// Comma-separated authorization scopes.
-    #[arg(long, value_name = "SCOPES")]
-    scopes: Option<String>,
-
-    /// Read defaults from this config file.
-    #[arg(long, short = 'c', value_name = "PATH")]
-    config: Option<String>,
-
-    /// Accept GoDaddy Developer agreements non-interactively when onboarding
-    /// is still pending (required for non-TTY).
-    #[arg(long)]
-    accept_agreements: bool,
-}
-
-/// `filesWritten` is a small path-by-kind object (`config`/`env`), so it
-/// renders as an indented property bag instead of a raw JSON dump.
-fn init_view_columns() -> Vec<TableColumn> {
-    vec![
-        TableColumn::new("id", "ID"),
-        TableColumn::new("name", "Name"),
-        TableColumn::new("status", "Status"),
-        TableColumn::new("clientId", "Client ID"),
-        TableColumn::new("orgId", "Org ID"),
-        TableColumn::new("url", "URL").no_truncate(true),
-        TableColumn::new("proxyUrl", "Proxy URL").no_truncate(true),
-        TableColumn::new("authorizationScopes", "Authorization Scopes"),
-        TableColumn::new("oauthGrantTypes", "OAuth Grant Types"),
-        TableColumn::new("filesWritten", "Files Written").nested(vec![
-            TableColumn::new("config", "Config").no_truncate(true),
-            TableColumn::new("env", "Env").no_truncate(true),
-        ]),
-    ]
-}
-
-fn init_command() -> RuntimeCommandSpec {
-    RuntimeCommandSpec::new_typed_with_context::<InitArgs, _, _, _>(
-        CommandSpec::from_args::<InitArgs>("init", "Create and initialize a new application")
-            .with_long(
-                "Register a new GoDaddy developer-platform application and write a \
-                godaddy.toml manifest to the current directory. The manifest captures the \
-                application name, URL, authorization scopes, and any actions or extensions \
-                added later. Run `gddy platform app validate <name>` to confirm the remote \
-                application is healthy, then `gddy platform app release` to create a versioned \
-                release.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Mutate)
-            .with_scopes(&[APP_REGISTRY_READ, APP_REGISTRY_WRITE])
-            .with_output_schema::<ApplicationInit>()
-            .with_view(init_view_columns()),
-        |ctx, args: InitArgs| async move {
-            let env = ctx.middleware.env.clone();
-            let config_path = crate::config::config_path(Some(&env));
-            let accept_agreements = args.accept_agreements;
-
-            // Seed defaults only from an explicit --config; a bad/missing --config is fatal.
-            let existing = match args.config.as_deref() {
-                Some(path) => Some(
-                    crate::config::read_config(std::path::Path::new(path)).map_err(|e| {
-                        crate::error::GddyError::config(format!("invalid config: {e}"))
-                            .into_cli_error()
-                    })?,
-                ),
-                None => None,
-            };
-
-            // Resolve each field: CLI flag, else the existing config's value, else a default.
-            let name = args
-                .name
-                .or_else(|| existing.as_ref().map(|c| c.name.clone()))
-                .unwrap_or_default();
-            let description = args
-                .description
-                .or_else(|| existing.as_ref().and_then(|c| c.description.clone()))
-                .unwrap_or_default();
-            let url = args
-                .url
-                .or_else(|| existing.as_ref().map(|c| c.url.clone()))
-                .unwrap_or_default();
-            let proxy_url = args
-                .proxy_url
-                .or_else(|| existing.as_ref().map(|c| c.proxy_url.clone()))
-                .unwrap_or_default();
-            let scopes: Vec<String> = args
-                .scopes
-                .map(|s| {
-                    s.split(',')
-                        .map(|p| p.trim())
-                        .filter(|p| !p.is_empty())
-                        .map(str::to_owned)
-                        .collect()
-                })
-                .or_else(|| existing.as_ref().map(|c| c.authorization_scopes.clone()))
-                .unwrap_or_default();
-            let label = args.label.unwrap_or_else(|| name.clone());
-
-            for (message, empty) in [
-                ("Application name is required", name.is_empty()),
-                (
-                    "Application description is required",
-                    description.is_empty(),
-                ),
-                ("Application URL is required", url.is_empty()),
-                ("Proxy URL is required", proxy_url.is_empty()),
-                ("Authorization scopes are required", scopes.is_empty()),
-            ] {
-                if empty {
-                    return Err(cli_engine::CliCoreError::message(message));
-                }
-            }
-
-            // Reject names that cannot be written to a valid godaddy.toml.
-            if !crate::config::is_valid_app_name(&name) {
-                return Err(crate::error::GddyError::validation(format!(
-                    "Application name must be 3-255 lowercase letters, digits, or hyphens \
-                     (got {name:?})"
-                ))
-                .into_cli_error());
-            }
-
-            for (field, u) in [("url", &url), ("proxyUrl", &proxy_url)] {
-                if !crate::application::public_url::is_public_routable_url(u) {
-                    return Err(cli_engine::CliCoreError::message(format!(
-                        "Invalid application configuration: {field} must be a publicly-resolvable \
-                         http(s) URL (localhost, loopback, and private IPs are not allowed)"
-                    )));
-                }
-            }
-
-            let credential = ctx.credential().await?;
-            let onboarding = crate::onboarding::ensure_ready_for_app_init(
-                &credential.token,
-                &env,
-                accept_agreements,
-            )
-            .await?;
-
-            let client = make_client(&ctx).await?;
-            let data = client
-                .create_application(json!({
-                    "name": name,
-                    "label": label,
-                    "description": description,
-                    "url": url,
-                    "proxyUrl": proxy_url,
-                    "organizationId": &onboarding.org_id,
-                    "authorizationScopes": scopes,
-                }))
-                .await
-                .map_err(client_err)?;
-
-            let app = &data["createApplication"];
-            // Credentials/secrets the API returns (all selected by create_application).
-            let client_id = app["clientId"].as_str().unwrap_or("").to_owned();
-            let client_secret = app["clientSecret"].as_str().unwrap_or("").to_owned();
-            let secret = app["secret"].as_str().unwrap_or("").to_owned();
-            let public_key = app["publicKey"].as_str().unwrap_or("").to_owned();
-
-            // Best-effort local writes: app create already succeeded. Only paths that
-            // actually write are included in filesWritten.
-            let config = crate::config::Config {
-                name: name.clone(),
-                client_id: client_id.clone(),
-                description: Some(description.clone()),
-                version: "0.0.0".to_owned(),
-                url: url.clone(),
-                proxy_url: proxy_url.clone(),
-                authorization_scopes: scopes.clone(),
-                actions: vec![],
-                subscriptions: Some(crate::config::SubscriptionsConfig { webhook: vec![] }),
-                dependencies: vec![],
-                extensions: None,
-            };
-            let cwd = match std::env::current_dir() {
-                Ok(dir) => dir,
-                Err(e) => {
-                    tracing::debug!(
-                        error = %e,
-                        "current_dir failed; filesWritten paths will be relative"
-                    );
-                    std::path::PathBuf::new()
-                }
-            };
-
-            let mut files_written = serde_json::Map::new();
-            if let Err(e) = crate::config::write_config(&config_path, &config) {
-                tracing::warn!(error = %e, path = %config_path.display(), "failed to write config; continuing");
-            } else {
-                files_written.insert(
-                    "config".to_owned(),
-                    json!(cwd.join(&config_path).display().to_string()),
-                );
-            }
-
-            let env_file_path = crate::config::env_path(Some(&env));
-            if let Err(e) = crate::config::write_env_file(
-                Some(&env),
-                &secret,
-                &public_key,
-                &client_id,
-                &client_secret,
-            ) {
-                tracing::warn!(error = %e, path = %env_file_path.display(), "failed to write env file; continuing");
-            } else {
-                files_written.insert(
-                    "env".to_owned(),
-                    json!(cwd.join(&env_file_path).display().to_string()),
-                );
-            }
-
-            let app_id = app["id"].as_str().unwrap_or("").to_owned();
-            let result = json!({
-                "id": app_id,
-                "name": name,
-                "status": app["status"].as_str().unwrap_or("").to_owned(),
-                "clientId": client_id,
-                "orgId": onboarding.org_id,
-                "url": url,
-                "proxyUrl": proxy_url,
-                "authorizationScopes": scopes,
-                "oauthGrantTypes": ["authorization_code", "client_credentials"],
-                "filesWritten": files_written,
-            });
-            Ok(CommandResult::new(result).with_next_actions(vec![
-                next_action(
-                    "platform app add action --name <name> --url <url>",
-                    "Add first action",
-                ),
-                next_action(
-                    "platform app add subscription --name <name> --events <events> --url <url>",
-                    "Add webhook subscription",
-                ),
-                next_action(
-                    "platform app validate <name>",
-                    "Validate the remote application state",
-                )
-                .with_param("name", required_value(&name)),
-                next_action(
-                    "platform app release --application-id <application-id> --version <version>",
-                    "Create the first release",
-                )
-                .with_param("application-id", required_value(&app_id))
-                .with_param("version", NextActionParam::required()),
-            ]))
-        },
-    )
-}
-
-/// Missing URL is an error; missing proxy URL or INACTIVE status are warnings.
-fn validate_remote_application(app: &Value) -> (bool, Vec<String>, Vec<String>) {
-    let mut errors = Vec::new();
-    let mut warnings = Vec::new();
-
-    let url = app["url"].as_str().unwrap_or("");
-    if url.is_empty() {
-        errors.push("Application URL is required".to_owned());
-    }
-    let proxy_url = app["proxyUrl"].as_str().unwrap_or("");
-    if proxy_url.is_empty() {
-        warnings.push("Proxy URL is not set".to_owned());
-    }
-    if app["status"].as_str() == Some("INACTIVE") {
-        warnings.push("Application is currently inactive".to_owned());
-    }
-
-    (errors.is_empty(), errors, warnings)
-}
-
-#[derive(Debug, Clone, clap::Args)]
-struct ValidateArgs {
-    /// Application name.
-    #[arg(value_name = "NAME")]
-    name: String,
-}
-
-fn validate_command() -> RuntimeCommandSpec {
-    RuntimeCommandSpec::new_typed_with_context::<ValidateArgs, _, _, _>(
-        CommandSpec::from_args::<ValidateArgs>("validate", "Validate remote application state")
-            .with_long(
-                "Fetch a GoDaddy developer-platform application by name and validate its \
-                remote configuration. Reports an error when the application URL is missing, \
-                and warnings when the proxy URL is unset or the application is inactive. \
-                Requires authentication.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Read)
-            .with_output_schema::<ValidationResult>(),
-        |ctx, args: ValidateArgs| async move {
-            let name = args.name;
-            let client = make_client(&ctx).await?;
-            let data = client.get_application(&name).await.map_err(client_err)?;
-            let app = &data["application"];
-            if app.is_null() {
-                return Err(crate::error::GddyError::not_found(format!(
-                    "application '{name}' not found"
-                ))
-                .into_cli_error());
-            }
-
-            let app_id = app["id"].as_str().unwrap_or("").to_owned();
-            let (valid, errors, warnings) = validate_remote_application(app);
-
-            // Only suggest a release once the app is valid; otherwise point back to
-            // `info` to review the reported problems.
-            let mut next_actions = Vec::new();
-            if valid {
-                next_actions.push(
-                    next_action(
-                        "platform app release --application-id <application-id> --version <version>",
-                        "Create a release after validation",
-                    )
-                    .with_param("application-id", required_value(&app_id))
-                    .with_param("version", NextActionParam::required()),
-                );
-            }
-            next_actions.push(
-                next_action(
-                    "platform app info --name <name>",
-                    "Inspect application details",
-                )
-                .with_param("name", required_value(&name)),
-            );
-
-            Ok(CommandResult::new(json!({
-                "valid": valid,
-                "errors": errors,
-                "warnings": warnings,
-            }))
-            .with_next_actions(next_actions))
-        },
-    )
-}
-
-/// "At least one of" fields for `update`, flattened into [`UpdateArgs`].
-///
-/// Kept as its own derive struct (rather than inline fields on `UpdateArgs`)
-/// so the struct-level `#[group(...)]` only covers these three — `id` stays
-/// outside the group and independently required. See the flatten caveat on
-/// `CommandSpec::from_args`: a struct can't both flatten a field and declare
-/// its own enforced group.
-#[derive(Debug, Clone, clap::Args)]
-#[group(required = true, multiple = true)]
-struct UpdateFields {
-    /// New label.
-    #[arg(long, value_name = "LABEL")]
-    label: Option<String>,
-
-    /// New description.
-    #[arg(long, value_name = "TEXT")]
-    description: Option<String>,
-
-    /// Application status (ACTIVE or INACTIVE).
-    #[arg(long, value_name = "STATUS", value_parser = ["ACTIVE", "INACTIVE"])]
-    status: Option<String>,
-}
-
-#[derive(Debug, Clone, clap::Args)]
-struct UpdateArgs {
-    /// Application ID.
-    #[arg(long, value_name = "ID")]
-    id: String,
-
-    #[command(flatten)]
-    fields: UpdateFields,
-}
-
-fn update_command() -> RuntimeCommandSpec {
-    RuntimeCommandSpec::new_typed_with_context::<UpdateArgs, _, _, _>(
-        CommandSpec::from_args::<UpdateArgs>("update", "Update an application")
-            .with_long(
-                "Update the label, description, or status of a GoDaddy developer-platform \
-                application by its ID. At least one of --label, --description, or --status \
-                must be provided. Use `gddy platform app info --name <name>` to retrieve the \
-                application ID.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Mutate)
-            .with_scopes(&[APP_REGISTRY_READ, APP_REGISTRY_WRITE])
-            .with_output_schema::<ApplicationUpdate>(),
-        |context, args: UpdateArgs| async move {
-            let mut input = serde_json::Map::new();
-            if let Some(label) = args.fields.label {
-                input.insert("label".to_owned(), json!(label));
-            }
-            if let Some(description) = args.fields.description {
-                input.insert("description".to_owned(), json!(description));
-            }
-            if let Some(status) = args.fields.status {
-                input.insert("status".to_owned(), json!(status));
-            }
-            let client = make_client(&context).await?;
-            let data = client
-                .update_application(&args.id, json!(input))
-                .await
-                .map_err(client_err)?;
-            let name = data["updateApplication"]["name"]
-                .as_str()
-                .unwrap_or("")
-                .to_owned();
-            Ok(
-                CommandResult::new(data["updateApplication"].clone()).with_next_actions(vec![
-                    next_action(
-                        "platform app info --name <name>",
-                        "Inspect updated application",
-                    )
-                    .with_param("name", required_value(&name)),
-                    next_action(
-                        "platform app deploy --name <name>",
-                        "Deploy updated application",
-                    )
-                    .with_param("name", required_value(&name)),
-                ]),
-            )
-        },
-    )
-}
-
-#[derive(Debug, Clone, clap::Args)]
-struct EnableDisableArgs {
-    /// Application name.
-    #[arg(value_name = "NAME")]
-    name: String,
-
-    /// Store ID to enable/disable the application on.
-    #[arg(long = "store-id", value_name = "STORE_ID")]
-    store_id: String,
-}
-
-fn enable_command() -> RuntimeCommandSpec {
-    RuntimeCommandSpec::new_typed_with_context::<EnableDisableArgs, _, _, _>(
-        CommandSpec::from_args::<EnableDisableArgs>("enable", "Enable an application on a store")
-            .with_long(
-                "Make a GoDaddy developer-platform application available on a specific \
-                store. Use `gddy platform app disable` to reverse this. Both the application name \
-                and a store ID are required.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Mutate)
-            .with_scopes(&[APP_REGISTRY_READ, APP_REGISTRY_WRITE])
-            .with_output_schema::<ApplicationRef>(),
-        |ctx, args: EnableDisableArgs| async move {
-            let name = args.name;
-            let store_id = args.store_id;
-            let client = make_client(&ctx).await?;
-            let data = client
-                .enable_application(json!({ "applicationName": name, "storeId": store_id }))
-                .await
-                .map_err(client_err)?;
-            Ok(
-                CommandResult::new(data["enableStoreApplication"].clone()).with_next_actions(vec![
-                    next_action(
-                        "platform app disable <name> --store-id <store-id>",
-                        "Disable the application on the same store",
-                    )
-                    .with_param("name", required_value(&name))
-                    .with_param("store-id", required_value(&store_id)),
-                    next_action(
-                        "platform app info --name <name>",
-                        "Inspect application status",
-                    )
-                    .with_param("name", required_value(&name)),
-                ]),
-            )
-        },
-    )
-}
-
-fn disable_command() -> RuntimeCommandSpec {
-    RuntimeCommandSpec::new_typed_with_context::<EnableDisableArgs, _, _, _>(
-        CommandSpec::from_args::<EnableDisableArgs>("disable", "Disable an application on a store")
-            .with_long(
-                "Remove a GoDaddy developer-platform application from a specific store, \
-                making it unavailable there. Use `gddy platform app enable` to re-enable it. \
-                Both the application name and a store ID are required.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Mutate)
-            .with_scopes(&[APP_REGISTRY_READ, APP_REGISTRY_WRITE])
-            .with_output_schema::<ApplicationRef>(),
-        |ctx, args: EnableDisableArgs| async move {
-            let name = args.name;
-            let store_id = args.store_id;
-            let client = make_client(&ctx).await?;
-            let data = client
-                .disable_application(json!({ "applicationName": name, "storeId": store_id }))
-                .await
-                .map_err(client_err)?;
-            Ok(
-                CommandResult::new(data["disableStoreApplication"].clone()).with_next_actions(
-                    vec![
-                        next_action(
-                            "platform app enable <name> --store-id <store-id>",
-                            "Re-enable the application on the same store",
-                        )
-                        .with_param("name", required_value(&name))
-                        .with_param("store-id", required_value(&store_id)),
-                        next_action(
-                            "platform app info --name <name>",
-                            "Inspect application status",
-                        )
-                        .with_param("name", required_value(&name)),
-                    ],
-                ),
-            )
-        },
-    )
-}
-
-fn archive_command() -> RuntimeCommandSpec {
-    RuntimeCommandSpec::new_typed_with_context::<ValidateArgs, _, _, _>(
-        CommandSpec::from_args::<ValidateArgs>("archive", "Archive an application")
-            .with_long(
-                "Archive a GoDaddy developer-platform application by name. Archiving is \
-                irreversible via this CLI; the application will no longer be active. \
-                Use `gddy platform app list` to confirm the application name before archiving.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Destructive)
-            .with_scopes(&[APP_REGISTRY_READ, APP_REGISTRY_WRITE])
-            .with_output_schema::<ApplicationArchive>(),
-        |ctx, args: ValidateArgs| async move {
-            let name = args.name;
-            let client = make_client(&ctx).await?;
-            let app_data = client.get_application(&name).await.map_err(client_err)?;
-            let app_id = app_data["application"]["id"]
-                .as_str()
-                .ok_or_else(|| {
-                    crate::error::GddyError::not_found(format!("application '{name}' not found"))
-                        .into_cli_error()
-                })?
-                .to_owned();
-            let data = client
-                .archive_application(&app_id)
-                .await
-                .map_err(client_err)?;
-            Ok(
-                CommandResult::new(data["archiveApplication"].clone()).with_next_actions(vec![
-                    next_action(
-                        "platform app info --name <name>",
-                        "Inspect archived application",
-                    )
-                    .with_param("name", required_value(&name)),
-                    next_action("platform app list", "List all platform apps"),
-                ]),
-            )
-        },
-    )
-}
-
-/// Build one `uiExtensions` release entry, enforcing the API's one-target-per-
-/// extension limit. `target` is omitted when the extension has no targets.
-fn ui_extension_entry(
-    name: &str,
-    handle: &str,
-    source: &str,
-    kind: &str,
-    targets: &[crate::config::ExtensionTarget],
-) -> cli_engine::Result<Value> {
-    if targets.len() > 1 {
-        return Err(cli_engine::CliCoreError::message(format!(
-            "UI extension '{name}' has {} targets, but only one target is supported per extension during release",
-            targets.len()
-        )));
-    }
-    let mut entry = json!({ "name": name, "handle": handle, "source": source, "type": kind });
-    if let Some(t) = targets.first() {
-        entry["target"] = json!(t.target);
-    }
-    Ok(entry)
-}
-
-/// Map godaddy.toml extensions (embed / checkout / blocks) to the release
-/// `uiExtensions` input. Mirrors the TS release mapping (single target each).
-fn build_ui_extensions(config: &crate::config::Config) -> cli_engine::Result<Vec<Value>> {
-    let mut out = Vec::new();
-    let Some(exts) = &config.extensions else {
-        return Ok(out);
-    };
-    for e in &exts.embed {
-        out.push(ui_extension_entry(
-            &e.name, &e.handle, &e.source, "embed", &e.targets,
-        )?);
-    }
-    for e in &exts.checkout {
-        out.push(ui_extension_entry(
-            &e.name, &e.handle, &e.source, "checkout", &e.targets,
-        )?);
-    }
-    if let Some(b) = &exts.blocks {
-        // Blocks carries no name/handle/targets in config; use the same fixed
-        // identifiers the TS release path uses.
-        out.push(
-            json!({ "name": "Blocks", "handle": "blocks", "source": b.source, "type": "blocks" }),
-        );
-    }
-    Ok(out)
-}
-
-#[derive(Debug, Clone, clap::Args)]
-struct ReleaseArgs {
-    /// Application ID.
-    #[arg(long = "application-id", value_name = "ID")]
-    application_id: String,
-
-    /// Semver release version.
-    #[arg(long, value_name = "VERSION")]
-    version: String,
-
-    /// Release description.
-    #[arg(long, value_name = "TEXT")]
-    description: Option<String>,
-}
-
-fn release_command() -> RuntimeCommandSpec {
-    RuntimeCommandSpec::new_typed_with_context::<ReleaseArgs, _, _, _>(
-        CommandSpec::from_args::<ReleaseArgs>("release", "Create a new application release")
-            .with_long(
-                "Tag a new versioned release for a GoDaddy developer-platform application. \
-                The version must follow semver (e.g. 1.2.3). A release is required before \
-                running `gddy platform app deploy`. Use `gddy platform app info --name <name>` to \
-                retrieve the application ID.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Mutate)
-            .with_scopes(&[APP_REGISTRY_READ, APP_REGISTRY_WRITE])
-            .with_output_schema::<ApplicationRelease>(),
-        |ctx, args: ReleaseArgs| async move {
-            let app_id = args.application_id;
-            let version = args.version;
-            let description = args.description;
-            let mut input = json!({ "applicationId": app_id, "version": version });
-            if let Some(desc) = description {
-                input["description"] = json!(desc);
-            }
-
-            let config_path = crate::config::config_path(Some(&ctx.middleware.env));
-            // Include actions, webhook subscriptions, and UI extensions from
-            // godaddy.toml so configured behavior is captured in the release.
-            // Without this, everything added via `platform app add` was silently
-            // dropped. A missing or invalid config is non-fatal (empty arrays);
-            // too many targets per extension is a hard error.
-            let (actions, subscriptions, ui_extensions) = match crate::config::read_config(
-                &config_path,
-            ) {
-                Ok(config) => {
-                    let actions: Vec<Value> = config
-                        .actions
-                        .iter()
-                        .map(|a| json!({ "name": a.name, "url": a.url }))
-                        .collect();
-                    let subscriptions: Vec<Value> = config
-                        .subscriptions
-                        .as_ref()
-                        .map(|s| {
-                            s.webhook
-                                .iter()
-                                .map(
-                                    |w| json!({ "name": w.name, "events": w.events, "url": w.url }),
-                                )
-                                .collect()
-                        })
-                        .unwrap_or_default();
-                    let ui_extensions = build_ui_extensions(&config)?;
-                    (actions, subscriptions, ui_extensions)
-                }
-                Err(e) => {
-                    tracing::warn!(
-                        error = %e,
-                        path = %config_path.display(),
-                        "failed to read config; releasing with empty actions, subscriptions, and uiExtensions"
-                    );
-                    (Vec::new(), Vec::new(), Vec::new())
-                }
-            };
-            input["actions"] = json!(actions);
-            input["subscriptions"] = json!(subscriptions);
-            input["uiExtensions"] = json!(ui_extensions);
-
-            let client = make_client(&ctx).await?;
-            let data = client.create_release(input).await.map_err(client_err)?;
-            // Release is keyed by `--application-id`, not name. Do not prefill
-            // `name` from godaddy.toml — that manifest may belong to a different app.
-            let name_param = NextActionParam::required();
-            Ok(
-                CommandResult::new(data["createRelease"].clone()).with_next_actions(vec![
-                    next_action(
-                        "platform app deploy --name <name>",
-                        "Deploy the released application",
-                    )
-                    .with_param("name", name_param.clone()),
-                    next_action(
-                        "platform app info --name <name>",
-                        "Inspect application and latest release",
-                    )
-                    .with_param("name", name_param),
-                ]),
-            )
-        },
-    )
-}
-
-#[derive(Debug, Clone, clap::Args)]
-struct DeployArgs {
-    /// Application name.
-    #[arg(long, short = 'n', value_name = "NAME")]
-    name: String,
-}
-
-fn deploy_command() -> RuntimeCommandSpec {
-    RuntimeCommandSpec::new_typed_streaming::<DeployArgs, _, _>(
-        CommandSpec::from_args::<DeployArgs>(
-            "deploy",
-            "Deploy an application, streaming progress events",
-        )
-        .with_long(
-            "Read godaddy.toml from the current directory, bundle all declared extensions \
-                with esbuild, run the security scanner (rules SEC101–SEC115) on each bundle, \
-                then upload the artifacts to the latest release of the named application. \
-                Progress is streamed as JSON events. A release must exist before deploying; \
-                create one with `gddy platform app release`.",
-        )
-        .with_system("applications")
-        .with_tier(Tier::Mutate)
-        .with_scopes(&[APP_REGISTRY_READ, APP_REGISTRY_WRITE]),
-        |ctx, args: DeployArgs, sender: StreamSender| async move {
-            let name = args.name;
-            let env = ctx.middleware.env.clone();
-            let token = tap_deploy_err(&sender, ctx.credential().await).await?.token;
-            let base_url = tap_deploy_err(&sender, api_url_for_env(&env)).await?;
-            let client = ApplicationClient::new(base_url, token);
-
-            sender
-                .send(json!({ "type": "step", "name": "auth.check", "status": "completed" }))
-                .await;
-
-            // Read the local godaddy.toml
-            sender
-                .send(json!({ "type": "step", "name": "config.read", "status": "started" }))
-                .await;
-            let config_path = crate::config::config_path(Some(&env));
-            let config = tap_deploy_err(
-                &sender,
-                crate::config::read_config(&config_path).map_err(|e| {
-                    crate::error::GddyError::config(format!("config read failed: {e}"))
-                        .into_cli_error()
-                }),
-            )
-            .await?;
-            sender
-                .send(json!({ "type": "step", "name": "config.read", "status": "completed", "path": config_path.display().to_string() }))
-                .await;
-
-            // Look up the application and its latest release
-            sender
-                .send(json!({ "type": "step", "name": "application.lookup", "status": "started" }))
-                .await;
-            let app_data = tap_deploy_err(
-                &sender,
-                client
-                    .get_application_with_releases(&name)
-                    .await
-                    .map_err(client_err),
-            )
-            .await?;
-            let app = &app_data["application"];
-            if app.is_null() {
-                let err =
-                    crate::error::GddyError::not_found(format!("application '{name}' not found"))
-                        .into_cli_error();
-                return Err(fail_deploy(&sender, err).await);
-            }
-            let application_id = app["id"].as_str().unwrap_or("").to_owned();
-            sender
-                .send(json!({ "type": "step", "name": "application.lookup", "status": "completed", "id": application_id }))
-                .await;
-
-            // Resolve latest release
-            sender
-                .send(json!({ "type": "step", "name": "release.lookup", "status": "started" }))
-                .await;
-            let release_id = app["releases"]["edges"]
-                .as_array()
-                .and_then(|edges| edges.first())
-                .and_then(|e| e["node"]["id"].as_str())
-                .map(str::to_owned);
-            let Some(release_id) = release_id else {
-                let err = cli_engine::CliCoreError::message(format!(
-                    "application '{name}' has no releases — create one first with: gddy platform app release --application-id {application_id} --version 0.0.1"
-                ));
-                return Err(fail_deploy(&sender, err).await);
-            };
-            sender
-                .send(json!({ "type": "step", "name": "release.lookup", "status": "completed", "releaseId": release_id }))
-                .await;
-
-            // Process each extension
-            let extensions = collect_extensions(&config);
-            let total = extensions.len();
-            // One timestamp for the whole deploy so all artifacts share a temp root.
-            let deploy_timestamp = crate::extension::format_timestamp(chrono::Utc::now());
-            sender
-                .send(json!({ "type": "step", "name": "extensions", "status": "started", "total": total }))
-                .await;
-
-            for (i, ext) in extensions.iter().enumerate() {
-                tap_deploy_err(
-                    &sender,
-                    deploy_extension(
-                        &client,
-                        &sender,
-                        DeployExtensionArgs {
-                            application_id: &application_id,
-                            release_id: &release_id,
-                            ext,
-                            index: i + 1,
-                            total,
-                            deploy_timestamp: &deploy_timestamp,
-                        },
-                    )
-                    .await,
-                )
-                .await?;
-            }
-
-            tap_deploy_err(
-                &sender,
-                finalize_deploy_activation(&client, &sender, &application_id, &release_id).await,
-            )
-            .await?;
-
-            sender
-                .send(json!({
-                    "type": "step",
-                    "name": "deploy",
-                    "status": "completed",
-                    "application": name,
-                    "extensions": total,
-                }))
-                .await;
-
-            sender
-                .send(deploy_result_event(
-                    &name,
-                    &application_id,
-                    &release_id,
-                    total,
-                ))
-                .await;
-            Ok(())
-        },
-    )
-}
-
-/// Finalize a deploy: activate the release, then promote the application to
-/// `ACTIVE`. Deploy must activate the release before promoting the application.
-async fn finalize_deploy_activation(
-    client: &ApplicationClient,
-    sender: &StreamSender,
-    application_id: &str,
-    release_id: &str,
-) -> cli_engine::Result<()> {
-    sender
-        .send(json!({ "type": "step", "name": "release.activate", "status": "started" }))
-        .await;
-    client
-        .activate_release(application_id, release_id)
-        .await
-        .map_err(client_err)?;
-    sender
-        .send(json!({ "type": "step", "name": "release.activate", "status": "completed" }))
-        .await;
-    sender
-        .send(json!({ "type": "step", "name": "application.activate", "status": "started" }))
-        .await;
-    client
-        .update_application(application_id, json!({ "status": "ACTIVE" }))
-        .await
-        .map_err(client_err)?;
-    sender
-        .send(json!({ "type": "step", "name": "application.activate", "status": "completed" }))
-        .await;
-
-    Ok(())
-}
-
-/// One extension entry from godaddy.toml ready for deploy: name, handle,
-/// source, type, and target surfaces (empty for blocks / untargeted).
-struct ExtensionDeploy {
-    name: String,
-    handle: String,
-    source: String,
-    ext_type: crate::extension::ExtensionType,
-    targets: Vec<String>,
-}
-
-fn collect_extensions(config: &crate::config::Config) -> Vec<ExtensionDeploy> {
-    let mut result = Vec::new();
-    if let Some(exts) = &config.extensions {
-        for e in &exts.embed {
-            result.push(ExtensionDeploy {
-                name: e.name.clone(),
-                handle: e.handle.clone(),
-                source: e.source.clone(),
-                ext_type: crate::extension::ExtensionType::Embed,
-                targets: e.targets.iter().map(|t| t.target.clone()).collect(),
-            });
-        }
-        for e in &exts.checkout {
-            result.push(ExtensionDeploy {
-                name: e.name.clone(),
-                handle: e.handle.clone(),
-                source: e.source.clone(),
-                ext_type: crate::extension::ExtensionType::Checkout,
-                targets: e.targets.iter().map(|t| t.target.clone()).collect(),
-            });
-        }
-        if let Some(blocks) = &exts.blocks {
-            result.push(ExtensionDeploy {
-                name: "Blocks".to_owned(),
-                handle: "blocks".to_owned(),
-                source: blocks.source.clone(),
-                ext_type: crate::extension::ExtensionType::Blocks,
-                targets: Vec::new(),
-            });
-        }
-    }
-    result
-}
-
-/// Resolve the upload target(s) for one extension. A blocks extension always
-/// uploads to the `blocks` target; embed/checkout upload once per configured
-/// target, or a single untargeted upload (`None`) when none are configured.
-fn resolve_upload_targets(
-    ext_type: crate::extension::ExtensionType,
-    targets: &[String],
-) -> Vec<Option<String>> {
-    match ext_type {
-        crate::extension::ExtensionType::Blocks => vec![Some("blocks".to_owned())],
-        _ if targets.is_empty() => vec![None],
-        _ => targets.iter().cloned().map(Some).collect(),
-    }
-}
-
-fn upload_completed_event(
-    extension_name: &str,
-    target: &Option<String>,
-    is_final_target: bool,
-    index: usize,
-    total: usize,
-) -> Value {
-    let mut event = json!({
-        "type": "progress",
-        "name": "extension.upload",
-        "status": "completed",
-        "extensionName": extension_name,
-        "target": target,
-    });
-    if is_final_target {
-        event["percent"] = json!(index * 100 / total.max(1));
-    }
-    event
-}
-
-/// Parse a comma-separated `--target` value into extension targets, trimming
-/// whitespace and dropping empty entries.
-fn parse_targets(raw: &str) -> Vec<crate::config::ExtensionTarget> {
-    raw.split(',')
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-        .map(|s| crate::config::ExtensionTarget {
-            target: s.to_owned(),
-        })
-        .collect()
-}
-
-/// Sandbox + normalize `--source` the same way deploy resolves it, so
-/// `godaddy.toml` always stores a path relative to `extensions/{handle}/`.
-fn normalized_extension_source(
-    handle: &str,
-    source: &str,
-    extension_name: &str,
-) -> cli_engine::Result<String> {
-    let repo_root = crate::extension::repo_root_from_cwd();
-    crate::extension::normalize_extension_source_for_config(
-        &repo_root,
-        handle,
-        source,
-        extension_name,
-    )
-    .map_err(validation_err)
-}
-
-fn validation_err(message: impl Into<String>) -> cli_engine::CliCoreError {
-    crate::error::GddyError::validation(message).into_cli_error()
-}
-
-struct DeployExtensionArgs<'a> {
-    application_id: &'a str,
-    release_id: &'a str,
-    ext: &'a ExtensionDeploy,
-    index: usize,
-    total: usize,
-    deploy_timestamp: &'a str,
-}
-
-async fn deploy_extension(
-    client: &ApplicationClient,
-    sender: &StreamSender,
-    args: DeployExtensionArgs<'_>,
-) -> cli_engine::Result<()> {
-    let DeployExtensionArgs {
-        application_id,
-        release_id,
-        ext,
-        index,
-        total,
-        deploy_timestamp,
-    } = args;
-    let ext_name = &ext.name;
-    let ext_type = ext.ext_type;
-
-    // ---- Bundle ----
-    sender
-        .send(json!({
-            "type": "progress",
-            "name": "extension.bundle",
-            "status": "started",
-            "extensionName": ext_name,
-            "percent": ((index - 1) * 100 / total.max(1)),
-        }))
-        .await;
-
-    let repo_root = crate::extension::repo_root_from_cwd();
-    let (ext_dir, source) =
-        crate::extension::resolve_extension_paths(&repo_root, &ext.handle, &ext.source, ext_name)
-            .map_err(validation_err)?;
-    crate::extension::require_extension_source_file(ext.handle.as_str(), ext_name, &source)
-        .map_err(validation_err)?;
-    let bundle = crate::extension::bundle_extension(
-        &source,
-        ext_type,
-        &ext_dir,
-        crate::extension::BundleOptions {
-            name: &ext.handle,
-            version: None,
-            repo_root: &repo_root,
-            timestamp: Some(deploy_timestamp),
-        },
-    )
-    .await
-    .map_err(|e| validation_err(format!("bundle failed for '{ext_name}': {e}")))?;
-    // Always clean temp artifacts when this function returns (success or error).
-    let _bundle_cleanup = crate::extension::BundleCleanup::new(&bundle);
-
-    sender
-        .send(json!({
-            "type": "progress",
-            "name": "extension.bundle",
-            "status": "completed",
-            "extensionName": ext_name,
-            "artifactName": bundle.artifact_name,
-            "artifactPath": bundle.artifact_path.display().to_string(),
-            "size": bundle.size,
-            "sha256": bundle.sha256,
-            "sourcemapPath": bundle.sourcemap_path.as_ref().map(|p| p.display().to_string()),
-        }))
-        .await;
-
-    // ---- Security scan ----
-    sender
-        .send(json!({
-            "type": "progress",
-            "name": "extension.scan",
-            "status": "started",
-            "extensionName": ext_name,
-            "artifactName": bundle.artifact_name,
-        }))
-        .await;
-
-    let source_display = ext.source.as_str();
-    let content = String::from_utf8_lossy(&bundle.bytes);
-    let findings = crate::extension::scan_bundle(&content, source_display);
-
-    if crate::extension::is_blocked(&findings) {
-        let blocked_msgs: Vec<String> = findings
-            .iter()
-            .filter(|f| f.severity == crate::extension::Severity::Block)
-            .map(|f| {
-                if f.snippet.is_empty() {
-                    format!("  {} ({}:{}): {}", f.rule_id, f.file, f.line, f.message)
-                } else {
-                    format!(
-                        "  {} ({}:{}): {}\n    > {}",
-                        f.rule_id, f.file, f.line, f.message, f.snippet
-                    )
-                }
-            })
-            .collect();
-        return Err(crate::error::GddyError::security(format!(
-            "security scan blocked deployment of '{ext_name}':\n{}",
-            blocked_msgs.join("\n")
-        ))
-        .into_cli_error());
-    }
-
-    sender
-        .send(json!({
-            "type": "progress",
-            "name": "extension.scan",
-            "status": "completed",
-            "extensionName": ext_name,
-            "findings": findings.len(),
-        }))
-        .await;
-
-    let bytes = bytes::Bytes::from(bundle.bytes);
-
-    // Upload the bundle once per configured target (blocks -> "blocks";
-    // embed/checkout -> each target, or a single untargeted upload).
-    let upload_targets = resolve_upload_targets(ext_type, &ext.targets);
-    for (target_index, target) in upload_targets.iter().enumerate() {
-        sender
-            .send(json!({ "type": "progress", "name": "extension.upload", "status": "started", "extensionName": ext_name, "target": target }))
-            .await;
-
-        let mut upload_input = json!({
-            "applicationId": application_id,
-            "releaseId": release_id,
-            "contentType": "JS",
-        });
-        if let Some(t) = target {
-            upload_input["target"] = json!(t);
-        }
-
-        let upload_data = client
-            .generate_upload_url(upload_input)
-            .await
-            .map_err(client_err)?;
-
-        let upload = &upload_data["generateReleaseUploadUrl"];
-        let upload_url = upload["url"].as_str().unwrap_or("").to_owned();
-        let upload_id = upload["uploadId"].as_str().unwrap_or("").to_owned();
-        let max_size_bytes = upload["maxSizeBytes"].as_u64();
-
-        // Parse required headers from ["key:value"] array
-        let mut headers = serde_json::Map::new();
-        if let Some(arr) = upload["requiredHeaders"].as_array() {
-            for h in arr {
-                if let Some(s) = h.as_str()
-                    && let Some((k, v)) = s.split_once(':')
-                {
-                    headers.insert(k.trim().to_owned(), json!(v.trim()));
-                }
-            }
-        }
-
-        client
-            .upload_artifact(
-                &upload_url,
-                &upload_id,
-                &json!(headers),
-                max_size_bytes,
-                bytes.clone(),
-                UploadOptions::default(),
-            )
-            .await
-            .map_err(client_err)?;
-
-        sender
-            .send(upload_completed_event(
-                ext_name,
-                target,
-                target_index + 1 == upload_targets.len(),
-                index,
-                total,
-            ))
-            .await;
-    }
-
-    Ok(())
-}
-
-#[derive(Debug, Clone, clap::Args)]
-struct ActionArgs {
-    /// Unique action name written into godaddy.toml.
-    #[arg(long)]
-    name: String,
-
-    /// Public HTTPS URL the platform will invoke for this action.
-    #[arg(long)]
-    url: String,
-}
-
-#[derive(Debug, Clone, clap::Args)]
-struct SubscriptionArgs {
-    /// Unique subscription name written into godaddy.toml.
-    #[arg(long)]
-    name: String,
-
-    /// Public HTTPS URL that will receive webhook POST requests.
-    #[arg(long)]
-    url: String,
-
-    /// One or more event types to subscribe to (run `gddy platform webhook
-    /// events` to list valid values).
-    #[arg(long, value_name = "EVENT", required = true, num_args = 1..)]
-    events: Vec<String>,
-}
-
-/// Shared shape for `add extension embed`/`add extension checkout` — both
-/// register a UI extension by name/handle/source/target.
-#[derive(Debug, Clone, clap::Args)]
-struct UiExtensionArgs {
-    /// Unique extension name written into godaddy.toml.
-    #[arg(long)]
-    name: String,
-
-    /// Platform handle used to identify this extension surface.
-    #[arg(long)]
-    handle: String,
-
-    /// Path to the JavaScript entry-point file, relative to
-    /// extensions/<handle>/ (e.g. src/index.ts), or a path already under
-    /// that directory.
-    #[arg(long)]
-    source: String,
-
-    /// Comma-separated target surface(s) for this extension.
-    #[arg(long)]
-    target: String,
-}
-
-#[derive(Debug, Clone, clap::Args)]
-struct BlocksArgs {
-    /// Path to the JavaScript entry-point file for the blocks extension,
-    /// relative to extensions/blocks/ (e.g. src/index.ts), or a path
-    /// already under that directory.
-    #[arg(long)]
-    source: String,
-}
-
-pub fn add_group() -> RuntimeGroupSpec {
-    RuntimeGroupSpec::new(
-        GroupSpec::new("add", "Add components to an application").with_long(
-            "Append actions, webhook subscriptions, or UI extensions to the \
-                godaddy.toml manifest in the current directory. These commands do not \
-                require authentication and do not contact the API; run \
-                `gddy platform app deploy` to publish the updated manifest.",
-        ),
-    )
-    .with_command(RuntimeCommandSpec::new_typed_with_context::<
-        ActionArgs,
-        _,
-        _,
-        _,
-    >(
-        CommandSpec::from_args::<ActionArgs>("action", "Add an action to godaddy.toml")
-            .with_long(
-                "Append an action entry to the godaddy.toml manifest in the current \
-                    directory. An action is an HTTP endpoint that the platform calls on \
-                    behalf of the application; it is identified by a name and a public \
-                    HTTPS URL. The manifest is updated in place; run \
-                    `gddy platform app validate <name>` to confirm remote application state.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Mutate)
-            .with_output_schema::<ConfigAction>()
-            .no_auth(true),
-        |ctx, args: ActionArgs| async move {
-            let name = args.name;
-            let url = args.url;
-            let path = crate::config::config_path(Some(&ctx.middleware.env));
-            let mut config = crate::config::read_config(&path)
-                .map_err(|e| cli_engine::CliCoreError::message(e.to_string()))?;
-            config.actions.push(crate::config::ActionConfig {
-                name: name.clone(),
-                url: url.clone(),
-            });
-            crate::config::write_config(&path, &config)
-                .map_err(|e| cli_engine::CliCoreError::message(e.to_string()))?;
-            Ok(CommandResult::new(json!({ "name": name, "url": url }))
-                .with_next_actions(add_config_next_actions(&config.name)))
-        },
-    ))
-    .with_command(RuntimeCommandSpec::new_typed_with_context::<
-        SubscriptionArgs,
-        _,
-        _,
-        _,
-    >(
-        CommandSpec::from_args::<SubscriptionArgs>(
-            "subscription",
-            "Add a webhook subscription to godaddy.toml",
-        )
-        .with_long(
-            "Append a webhook subscription entry to the godaddy.toml manifest in \
-                    the current directory. A subscription routes platform events to an HTTPS \
-                    endpoint. Provide one or more event types with --events; run \
-                    `gddy platform webhook events` to discover the full list of valid event types.",
-        )
-        .with_system("applications")
-        .with_tier(Tier::Mutate)
-        .with_output_schema::<ConfigSubscription>()
-        .no_auth(true),
-        |ctx, args: SubscriptionArgs| async move {
-            let name = args.name;
-            let url = args.url;
-            let events = args.events;
-            let path = crate::config::config_path(Some(&ctx.middleware.env));
-            let mut config = crate::config::read_config(&path)
-                .map_err(|e| cli_engine::CliCoreError::message(e.to_string()))?;
-            let subs = config
-                .subscriptions
-                .get_or_insert_with(|| crate::config::SubscriptionsConfig { webhook: vec![] });
-            subs.webhook.push(crate::config::SubscriptionConfig {
-                name: name.clone(),
-                events: events.clone(),
-                url: url.clone(),
-            });
-            crate::config::write_config(&path, &config)
-                .map_err(|e| cli_engine::CliCoreError::message(e.to_string()))?;
-            Ok(
-                CommandResult::new(json!({ "name": name, "url": url, "events": events }))
-                    .with_next_actions(add_config_next_actions(&config.name)),
-            )
-        },
-    ))
-    .with_group(add_extension_group())
-}
-
-pub fn add_extension_group() -> RuntimeGroupSpec {
-    RuntimeGroupSpec::new(
-        GroupSpec::new("extension", "Add an extension to godaddy.toml").with_long(
-            "Append a UI extension entry to the godaddy.toml manifest in the current \
-                directory. Extensions are JavaScript bundles that the platform renders \
-                inside store surfaces. Three types are supported: embed (inline widget), \
-                checkout (checkout-flow widget), and blocks (content blocks). Run \
-                `gddy platform app deploy` to bundle and upload the registered extensions.",
-        ),
-    )
-    .with_command(RuntimeCommandSpec::new_typed_with_context::<
-        UiExtensionArgs,
-        _,
-        _,
-        _,
-    >(
-        CommandSpec::from_args::<UiExtensionArgs>("embed", "Add an embed extension")
-            .with_long(
-                "Register an embed UI extension in godaddy.toml. An embed extension is a \
-                JavaScript bundle rendered as an inline widget inside a store surface. \
-                Provide a unique name, a platform handle, and the path to the source \
-                entry-point file. Run `gddy platform app deploy` to bundle and upload.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Mutate)
-            .with_output_schema::<ExtensionHandle>()
-            .no_auth(true),
-        |ctx, args: UiExtensionArgs| async move {
-            let name = args.name;
-            let handle = args.handle;
-            let source = normalized_extension_source(&handle, &args.source, &name)?;
-            let targets = parse_targets(&args.target);
-            if targets.is_empty() {
-                return Err(crate::error::GddyError::validation(
-                    "at least one --target is required (comma-separated)",
-                )
-                .into_cli_error());
-            }
-            let path = crate::config::config_path(Some(&ctx.middleware.env));
-            let mut config = crate::config::read_config(&path)
-                .map_err(|e| crate::error::GddyError::config(e.to_string()).into_cli_error())?;
-            let exts = config
-                .extensions
-                .get_or_insert_with(|| crate::config::ExtensionsConfig {
-                    embed: vec![],
-                    checkout: vec![],
-                    blocks: None,
-                });
-            exts.embed.push(crate::config::EmbedExtensionConfig {
-                name: name.clone(),
-                handle: handle.clone(),
-                source,
-                targets,
-            });
-            crate::config::write_config(&path, &config)
-                .map_err(|e| crate::error::GddyError::config(e.to_string()).into_cli_error())?;
-            Ok(
-                CommandResult::new(json!({ "name": name, "handle": handle, "type": "embed" }))
-                    .with_next_actions(add_config_next_actions(&config.name)),
-            )
-        },
-    ))
-    .with_command(RuntimeCommandSpec::new_typed_with_context::<
-        UiExtensionArgs,
-        _,
-        _,
-        _,
-    >(
-        CommandSpec::from_args::<UiExtensionArgs>("checkout", "Add a checkout extension")
-            .with_long(
-                "Register a checkout UI extension in godaddy.toml. A checkout extension is \
-                a JavaScript bundle rendered during the store checkout flow. Provide a \
-                unique name, a platform handle, and the path to the source entry-point \
-                file. Run `gddy platform app deploy` to bundle and upload.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Mutate)
-            .with_output_schema::<ExtensionHandle>()
-            .no_auth(true),
-        |ctx, args: UiExtensionArgs| async move {
-            let name = args.name;
-            let handle = args.handle;
-            let source = normalized_extension_source(&handle, &args.source, &name)?;
-            let targets = parse_targets(&args.target);
-            if targets.is_empty() {
-                return Err(crate::error::GddyError::validation(
-                    "at least one --target is required (comma-separated)",
-                )
-                .into_cli_error());
-            }
-            let path = crate::config::config_path(Some(&ctx.middleware.env));
-            let mut config = crate::config::read_config(&path)
-                .map_err(|e| crate::error::GddyError::config(e.to_string()).into_cli_error())?;
-            let exts = config
-                .extensions
-                .get_or_insert_with(|| crate::config::ExtensionsConfig {
-                    embed: vec![],
-                    checkout: vec![],
-                    blocks: None,
-                });
-            exts.checkout.push(crate::config::CheckoutExtensionConfig {
-                name: name.clone(),
-                handle: handle.clone(),
-                source,
-                targets,
-            });
-            crate::config::write_config(&path, &config)
-                .map_err(|e| crate::error::GddyError::config(e.to_string()).into_cli_error())?;
-            Ok(
-                CommandResult::new(json!({ "name": name, "handle": handle, "type": "checkout" }))
-                    .with_next_actions(add_config_next_actions(&config.name)),
-            )
-        },
-    ))
-    .with_command(RuntimeCommandSpec::new_typed_with_context::<
-        BlocksArgs,
-        _,
-        _,
-        _,
-    >(
-        CommandSpec::from_args::<BlocksArgs>("blocks", "Add a blocks extension")
-            .with_long(
-                "Register a blocks UI extension in godaddy.toml. A blocks extension is a \
-                JavaScript bundle that provides content blocks within a store. Only one \
-                blocks extension is supported per application; running this command again \
-                will overwrite the existing entry. Run `gddy platform app deploy` to bundle and \
-                upload.",
-            )
-            .with_system("applications")
-            .with_tier(Tier::Mutate)
-            .with_output_schema::<ExtensionBlocks>()
-            .no_auth(true),
-        |ctx, args: BlocksArgs| async move {
-            let source = normalized_extension_source("blocks", &args.source, "Blocks")?;
-            let path = crate::config::config_path(Some(&ctx.middleware.env));
-            let mut config = crate::config::read_config(&path)
-                .map_err(|e| crate::error::GddyError::config(e.to_string()).into_cli_error())?;
-            let exts = config
-                .extensions
-                .get_or_insert_with(|| crate::config::ExtensionsConfig {
-                    embed: vec![],
-                    checkout: vec![],
-                    blocks: None,
-                });
-            exts.blocks = Some(crate::config::BlocksExtensionConfig {
-                source: source.clone(),
-            });
-            crate::config::write_config(&path, &config)
-                .map_err(|e| crate::error::GddyError::config(e.to_string()).into_cli_error())?;
-            Ok(
-                CommandResult::new(json!({ "source": source, "type": "blocks" }))
-                    .with_next_actions(add_config_next_actions(&config.name)),
-            )
-        },
-    ))
+    .with_command(list::command())
+    .with_command(info::command())
+    .with_command(init::command())
+    .with_command(validate::command())
+    .with_command(update::command())
+    .with_command(lifecycle::enable_command())
+    .with_command(lifecycle::disable_command())
+    .with_command(lifecycle::archive_command())
+    .with_command(release::command())
+    .with_command(deploy::command())
+    .with_group(add::group())
 }
 
 #[cfg(test)]
 mod tests {
-    use cli_engine::{Cli, CliConfig, Stage};
-    use serde_json::json;
-
-    use super::{
-        add_config_next_actions, deploy_next_actions, init_view_columns, update_command,
-        validate_command, validate_remote_application,
-    };
-
     #[test]
     fn reused_next_action_helpers_have_expected_size() {
-        assert_eq!(add_config_next_actions("app").len(), 2);
-        assert_eq!(deploy_next_actions("app").len(), 3);
+        assert_eq!(super::add_config_next_actions("app").len(), 2);
+        assert_eq!(super::deploy::deploy_next_actions("app").len(), 3);
     }
 
     #[test]
     fn add_config_next_actions_skips_empty_name_prefill() {
-        let actions = add_config_next_actions("");
+        let actions = super::add_config_next_actions("");
         let name = &actions[0].params["name"];
         assert!(name.required);
         assert_eq!(name.value.as_deref(), None);
-    }
-
-    /// Proves `init_view_columns()` renders a `filesWritten` shaped like what
-    /// the `init` handler actually builds (`config`/`env` paths, confirmed by
-    /// inspection) as a nested property bag — a column/field name mismatch
-    /// would silently drop the write summary from human output and print a
-    /// raw JSON blob instead. Renders a hand-built envelope rather than
-    /// calling the handler, which would need a live app-registry API call.
-    #[test]
-    fn init_result_renders_files_written_as_a_nested_property_bag() {
-        let result = json!({
-            "id": "app-1",
-            "name": "demo",
-            "status": "ACTIVE",
-            "filesWritten": {
-                "config": "/home/user/project/godaddy.toml",
-                "env": "/home/user/project/.env",
-            },
-        });
-        let envelope = cli_engine::Envelope::success(result, "applications");
-        let rendered =
-            cli_engine::render_human_with_view(&envelope, Some(&init_view_columns()), "");
-        assert!(rendered.contains("Files Written:"), "{rendered}");
-        assert!(rendered.contains("godaddy.toml"), "{rendered}");
-        assert!(rendered.contains(".env"), "{rendered}");
-    }
-
-    fn update_clap_command() -> clap::Command {
-        update_command().spec.clap_command()
-    }
-
-    fn validate_clap_command() -> clap::Command {
-        validate_command().spec.clap_command()
-    }
-
-    #[test]
-    fn validate_requires_name() {
-        let err = validate_clap_command()
-            .try_get_matches_from(["validate"])
-            .expect_err("validate without name should be rejected");
-        assert_eq!(
-            err.kind(),
-            clap::error::ErrorKind::MissingRequiredArgument,
-            "expected MissingRequiredArgument, got: {err}"
-        );
-    }
-
-    #[test]
-    fn validate_accepts_positional_name() {
-        validate_clap_command()
-            .try_get_matches_from(["validate", "my-app"])
-            .expect("positional name should be accepted");
-    }
-
-    #[test]
-    fn validate_remote_healthy_app_is_valid() {
-        let (valid, errors, warnings) = validate_remote_application(&json!({
-            "url": "https://example.com",
-            "proxyUrl": "https://proxy.example.com",
-            "status": "ACTIVE",
-        }));
-        assert!(valid);
-        assert!(errors.is_empty());
-        assert!(warnings.is_empty());
-    }
-
-    #[test]
-    fn validate_remote_missing_url_is_error() {
-        let (valid, errors, warnings) = validate_remote_application(&json!({
-            "url": "",
-            "proxyUrl": "https://proxy.example.com",
-            "status": "ACTIVE",
-        }));
-        assert!(!valid);
-        assert_eq!(errors, vec!["Application URL is required".to_owned()]);
-        assert!(warnings.is_empty());
-    }
-
-    #[test]
-    fn validate_remote_missing_proxy_and_inactive_are_warnings() {
-        let (valid, errors, warnings) = validate_remote_application(&json!({
-            "url": "https://example.com",
-            "proxyUrl": null,
-            "status": "INACTIVE",
-        }));
-        assert!(valid, "warnings alone should not invalidate");
-        assert!(errors.is_empty());
-        assert_eq!(
-            warnings,
-            vec![
-                "Proxy URL is not set".to_owned(),
-                "Application is currently inactive".to_owned(),
-            ]
-        );
-    }
-
-    #[test]
-    fn status_rejects_values_outside_active_inactive() {
-        for bad in ["active", "DISABLED", "PENDING"] {
-            let err = update_clap_command()
-                .try_get_matches_from(["update", "--id", "app-1", "--status", bad])
-                .expect_err("invalid --status should be rejected");
-            assert_eq!(
-                err.kind(),
-                clap::error::ErrorKind::InvalidValue,
-                "--status {bad:?} should fail possible-value validation, got: {err}"
-            );
-        }
-    }
-
-    #[test]
-    fn status_accepts_active_and_inactive() {
-        for good in ["ACTIVE", "INACTIVE"] {
-            update_clap_command()
-                .try_get_matches_from(["update", "--id", "app-1", "--status", good])
-                .expect("ACTIVE|INACTIVE --status should be accepted");
-        }
-    }
-
-    #[test]
-    fn update_requires_at_least_one_field() {
-        let err = update_clap_command()
-            .try_get_matches_from(["update", "--id", "app-1"])
-            .expect_err("update with only --id should be rejected at parse time");
-        assert_eq!(
-            err.kind(),
-            clap::error::ErrorKind::MissingRequiredArgument,
-            "expected MissingRequiredArgument, got: {err}"
-        );
-    }
-
-    #[test]
-    fn update_accepts_any_single_field() {
-        update_clap_command()
-            .try_get_matches_from(["update", "--id", "app-1", "--label", "New"])
-            .expect("--label alone should be accepted");
-        update_clap_command()
-            .try_get_matches_from(["update", "--id", "app-1", "--description", "Desc"])
-            .expect("--description alone should be accepted");
-        update_clap_command()
-            .try_get_matches_from(["update", "--id", "app-1", "--status", "ACTIVE"])
-            .expect("--status alone should be accepted");
-    }
-
-    #[test]
-    fn update_accepts_multiple_fields() {
-        update_clap_command()
-            .try_get_matches_from([
-                "update",
-                "--id",
-                "app-1",
-                "--label",
-                "New",
-                "--description",
-                "Desc",
-                "--status",
-                "INACTIVE",
-            ])
-            .expect("multiple update fields should be allowed together");
-    }
-
-    /// API commands must stay fail-closed: `platform app list` calls the backend,
-    /// so it must require authentication. Built with **no auth provider
-    /// registered**, the engine's default `AuthRequirement::Required` must reject
-    /// it before the handler runs (no network call). This guards against someone
-    /// mistakenly marking an API command `no_auth(true)`, which would let it run
-    /// unauthenticated.
-    #[tokio::test]
-    async fn platform_app_list_requires_auth() {
-        let cli = Cli::new(
-            CliConfig::new("gddy", "GoDaddy developer CLI", "gddy")
-                .with_min_stage(Stage::Experimental)
-                .with_default_auth_provider("godaddy")
-                .with_module(crate::platform::module()),
-        );
-
-        let output = cli
-            .run(["gddy", "platform", "app", "list", "--output", "json"])
-            .await;
-
-        // The engine maps auth-resolution failures to exit code 2; together with
-        // the provider-named error below this proves the command was rejected at
-        // credential resolution, not by the handler hitting the network.
-        const AUTH_FAILURE_EXIT: i32 = 2;
-        assert_eq!(
-            output.exit_code, AUTH_FAILURE_EXIT,
-            "platform app list must fail closed at auth resolution, got: {}",
-            output.rendered
-        );
-        let json: serde_json::Value =
-            serde_json::from_str(&output.rendered).expect("valid json output");
-        let message = json["error"]["message"].as_str().unwrap_or_default();
-        assert!(
-            message.contains("provider"),
-            "expected an auth-provider resolution error, got: {}",
-            output.rendered
-        );
-    }
-
-    #[test]
-    fn parse_targets_trims_splits_and_drops_empties() {
-        let parsed = super::parse_targets(" a , b ,, c ");
-        let got: Vec<String> = parsed.into_iter().map(|t| t.target).collect();
-        assert_eq!(got, vec!["a", "b", "c"]);
-        assert!(super::parse_targets("   ").is_empty());
-        assert!(super::parse_targets("").is_empty());
-    }
-
-    #[test]
-    fn resolve_upload_targets_by_type() {
-        use crate::extension::ExtensionType;
-
-        // Blocks always uploads to the fixed "blocks" target.
-        assert_eq!(
-            super::resolve_upload_targets(ExtensionType::Blocks, &[]),
-            vec![Some("blocks".to_owned())]
-        );
-        // Embed/checkout with no targets: a single untargeted upload.
-        assert_eq!(
-            super::resolve_upload_targets(ExtensionType::Embed, &[]),
-            vec![None]
-        );
-        // Embed/checkout with targets: one upload per target.
-        assert_eq!(
-            super::resolve_upload_targets(
-                ExtensionType::Checkout,
-                &["a".to_owned(), "b".to_owned()]
-            ),
-            vec![Some("a".to_owned()), Some("b".to_owned())]
-        );
-    }
-
-    #[test]
-    fn final_target_completion_preserves_legacy_upload_event() {
-        let target = Some("admin.product.detail".to_owned());
-
-        let event = super::upload_completed_event("widget", &target, true, 1, 1);
-
-        assert_eq!(
-            event,
-            serde_json::json!({
-                "type": "progress",
-                "name": "extension.upload",
-                "status": "completed",
-                "extensionName": "widget",
-                "target": "admin.product.detail",
-                "percent": 100,
-            })
-        );
-    }
-
-    #[test]
-    fn ui_extension_entry_maps_fields_and_target() {
-        use crate::config::ExtensionTarget;
-
-        // No targets: `target` is omitted.
-        let none = super::ui_extension_entry("Widget", "widget", "src/w.ts", "embed", &[])
-            .expect("entry builds");
-        assert_eq!(none["name"], "Widget");
-        assert_eq!(none["handle"], "widget");
-        assert_eq!(none["type"], "embed");
-        assert_eq!(none["source"], "src/w.ts");
-        assert!(none.get("target").is_none());
-
-        // Exactly one target: `target` is set to that value.
-        let one = super::ui_extension_entry(
-            "Widget",
-            "widget",
-            "src/w.ts",
-            "checkout",
-            &[ExtensionTarget {
-                target: "checkout.block".to_owned(),
-            }],
-        )
-        .expect("entry builds");
-        assert_eq!(one["target"], "checkout.block");
-    }
-
-    #[test]
-    fn ui_extension_entry_rejects_multiple_targets() {
-        use crate::config::ExtensionTarget;
-        let targets = vec![
-            ExtensionTarget {
-                target: "a".to_owned(),
-            },
-            ExtensionTarget {
-                target: "b".to_owned(),
-            },
-        ];
-        let err = super::ui_extension_entry("Widget", "widget", "src/w.ts", "embed", &targets)
-            .expect_err("more than one target must be rejected");
-        assert!(
-            err.to_string().contains("only one target is supported"),
-            "unexpected error: {err}"
-        );
-    }
-
-    /// `deploy --follow` must end with exactly one terminal line. `StreamSender`
-    /// has no public constructor outside cli-engine, so the send itself can't
-    /// be unit-tested here — instead this locks down the shape of the payload
-    /// that gets sent, which is where the actual logic (error-code derivation)
-    /// lives.
-    #[test]
-    fn deploy_error_event_falls_back_to_error_code_for_a_plain_message() {
-        let err = cli_engine::CliCoreError::message("application 'foo' not found");
-        let event = super::deploy_error_event(&err);
-
-        assert_eq!(event["type"], "error");
-        assert_eq!(event["ok"], false);
-        assert_eq!(event["error"]["code"], "ERROR");
-        assert_eq!(event["error"]["message"], "application 'foo' not found");
-        assert!(event.get("fix").is_none());
-        assert_eq!(event["next_actions"], serde_json::json!([]));
-    }
-
-    #[test]
-    fn deploy_error_event_includes_fix_from_gddy_error() {
-        let err =
-            crate::error::GddyError::not_found("application 'foo' not found").into_cli_error();
-        let event = super::deploy_error_event(&err);
-
-        assert_eq!(event["error"]["code"], crate::error::codes::NOT_FOUND);
-        assert_eq!(event["error"]["message"], "application 'foo' not found");
-        assert!(
-            event["fix"]
-                .as_str()
-                .is_some_and(|f| f.contains("platform app list")),
-            "expected fix on streaming error event: {event}"
-        );
-    }
-
-    #[derive(Debug)]
-    struct CodedTestError;
-
-    impl std::fmt::Display for CodedTestError {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            write!(f, "upstream rejected the release")
-        }
-    }
-
-    impl std::error::Error for CodedTestError {}
-
-    impl cli_engine::DetailedError for CodedTestError {
-        fn error_code(&self) -> std::borrow::Cow<'static, str> {
-            "RELEASE_REJECTED".into()
-        }
-
-        fn error_system(&self) -> Option<std::borrow::Cow<'static, str>> {
-            Some("applications".into())
-        }
-
-        fn error_request_id(&self) -> Option<std::borrow::Cow<'static, str>> {
-            None
-        }
-
-        fn error_fix(&self) -> Option<std::borrow::Cow<'static, str>> {
-            Some("Check release status and retry.".into())
-        }
-    }
-
-    /// Proves `deploy_error_event` actually passes through a real error code
-    /// (via `build_error_envelope`) rather than always falling back — the
-    /// fallback-only case above wouldn't catch a regression that hardcoded
-    /// `"ERROR"`.
-    #[test]
-    fn deploy_error_event_passes_through_a_real_error_code() {
-        let err = cli_engine::CliCoreError::with_detailed_error(CodedTestError);
-        let event = super::deploy_error_event(&err);
-
-        assert_eq!(event["error"]["code"], "RELEASE_REJECTED");
-        assert_eq!(event["error"]["message"], "upstream rejected the release");
-        assert_eq!(event["fix"], "Check release status and retry.");
-    }
-
-    #[test]
-    fn deploy_result_event_is_ok_with_summary_fields() {
-        let event = super::deploy_result_event("my-app", "app-123", "rel-456", 2);
-
-        assert_eq!(event["type"], "result");
-        assert_eq!(event["ok"], true);
-        assert_eq!(event["result"]["application"], "my-app");
-        assert_eq!(event["result"]["applicationId"], "app-123");
-        assert_eq!(event["result"]["releaseId"], "rel-456");
-        assert_eq!(event["result"]["extensions"], 2);
-        assert_eq!(event["result"]["status"], "ACTIVE");
-        assert_eq!(
-            event["next_actions"].as_array().map(|a| a.len()),
-            Some(3),
-            "deploy should suggest enable, info, and redeploy: {event}"
-        );
-        assert_eq!(
-            event["next_actions"][1]["params"]["name"],
-            serde_json::json!({ "value": "my-app", "required": true }),
-            "deployed application name should prefill the next action: {event}"
-        );
-        assert!(
-            event["next_actions"][0]["command"]
-                .as_str()
-                .unwrap_or("")
-                .contains("platform app enable"),
-            "first next action should enable on a store: {event}"
-        );
     }
 }
